flake test fixed
